requirements.txt
Django==4.1.7
lxml==4.9.2
PyExecJS==1.5.1
requests==2.28.2
